Bug fix, simplify the mechanics of updating the board
there was a regression around f62c7a0f390f14dd95dfaff466d56128a6c5534e
that caused the state of the board to oscillate around certain events.

rather than keep track of which events are causing the board moves to
update, remove the synchronization routines and triggers and pass the
move to the playing boards. This way we do not fetch a board from the
server and the state is never overwritten and no extraneous "change"
singals are fired.